New Version! See Description.
We fixed animation, it doesn't jump to a random frame and actually animates when Idle. Animation now gives us a warning if we try and set an animation that doesn't exist.

Broke draw order, but not for long! We are going to try load more stuff as .ase, and maybe remove the need for all extra members in asset_ase right now. Polymorphism perhaps.